Data-Pipeline/requirements.txt
fastapi
uvicorn
firecrawl
pinecone
requests
python-dotenv 